Working on complete deprecation of html-composer
