+++ lib/responses.js
@@ -426,6 +426,19 @@ function expectHTML(fn, res) {
 }
 
 /**
+ * TODO: FINISH
+ * @param  {Function} fn  [description]
+ * @param  {[type]}   res [description]
+ * @return [type]         [description]
+ */
+function expectResponseType(fn, res) {
+  bluebird.try(fn).then(function (result) {
+    res.type(result.type);
+    res.send(result.output);
+  }).catch(handleError(res));
+}
+
+/**
  * @param {[{key: string, value: string}]} list
  * @returns {object}
  */
@@ -563,6 +576,7 @@ module.exports.onlyAcceptExtensions = onlyAcceptExtensions; // nice 404 for unsu
 module.exports.expectText = expectText;
 module.exports.expectJSON = expectJSON;
 module.exports.expectHTML = expectHTML;
+module.exports.expectResponseType = expectResponseType;
 
 // straight from DB
 module.exports.list = list;
