+++ lib/routes.js
@@ -12,6 +12,7 @@ const _ = require('lodash'),
   path = require('path'),
   siteService = require('./services/sites'),
   htmlComposer = require('./html-composer'),
+  render = require('./render'),
   responses = require('./responses'),
   files = require('./files'),
   references = require('./services/references'),
@@ -185,7 +186,7 @@ function addSiteController(router, site) {
 
     if (controller) {
       if (_.isFunction(controller)) {
-        controller(router, htmlComposer, site);
+        controller(router, render, site);
       }
 
       // things to remember from controller
