+++ lib/routes/components.js
@@ -7,8 +7,11 @@
 'use strict';
 
 const _ = require('lodash'),
+  // html = require('amphora-html'),
   responses = require('../responses'),
   htmlComposer = require('../html-composer'),
+  render = require('../render'),
+  composer = require('../services/composer'),
   files = require('../files'),
   queryStringOptions = ['ignore-data'],
   controller = require('../services/components'),
@@ -150,12 +153,15 @@ route = _.bindAll({
    * @returns {Function}
    */
   getExtension(req, res) {
-    switch (req.params.ext.toLowerCase()) {
-      case 'html':
-        return this.render(req, res);
-      case 'json':
-      default:
-        return this.getComposed(req, res);
+    // Check if the extension is included in the renderers
+    // that have been registered for the application
+    if (render.rendererExists(req.params.ext.toLowerCase())) {
+      this.render(req, res);
+    } else {
+      // Otherwise let's send back JSON as the response
+      return responses.expectJSON(function () {
+        return getComposed(req.uri, res.locals);
+      }, res);
     }
   },
 
@@ -198,8 +204,8 @@ route = _.bindAll({
   },
 
   render(req, res) {
-    responses.expectHTML(function () {
-      return htmlComposer.renderComponent(req.uri, res, _.pick(req.query, queryStringOptions));
+    responses.expectResponseType(function () {
+      return render.renderComponent(req, res, _.pick(req.query, queryStringOptions));
     }, res);
   }
 }, [
