+++ lib/routes/pages.js
@@ -9,6 +9,7 @@
 const _ = require('lodash'),
   responses = require('../responses'),
   htmlComposer = require('../html-composer'),
+  render = require('../render'),
   controller = require('../services/pages'),
   db = require('../services/db'),
   acceptedExtensions = {
@@ -23,6 +24,7 @@ const _ = require('lodash'),
  * @return {Promise}        composed data of root component and children
  */
 function getComposed(uri, locals) {
+  console.log(uri);
   return db.get(uri)
     .then(JSON.parse)
     // require here because otherwise it's a circular dependency (via html-composer)
@@ -70,16 +72,15 @@ let route = _.bindAll({
    * @returns {Promise}
    */
   extension(req, res) {
-    switch (req.params.ext.toLowerCase()) {
-      case 'html':
-        req.headers.accept = 'text/html';
-        this.render(req, res);
-        break;
-      case 'json':
-      default:
-        return responses.expectJSON(function () {
-          return getComposed(req.uri, res.locals);
-        }, res);
+    // Check if the extension is included in the renderers
+    // that have been registered for the application
+    if (render.rendererExists(req.params.ext.toLowerCase())) {
+      this.render(req, res);
+    } else {
+      // Otherwise let's send back JSON as the response
+      return responses.expectJSON(function () {
+        return getComposed(req.uri, res.locals);
+      }, res);
     }
   },
   /**
@@ -87,8 +88,8 @@ let route = _.bindAll({
    * @param {object} res
    */
   render(req, res) {
-    responses.expectHTML(function () {
-      return htmlComposer.renderPage(req.uri, res);
+    responses.expectResponseType(function () {
+      return render.renderPage(req.uri, res);
     }, res);
   }
 }, [
