+++ lib/routes/uris.js
@@ -14,6 +14,7 @@ const responses = require('../responses'),
  * @param {object} res
  */
 function getUriFromReference(req, res) {
+  console.log(1);
   responses.expectText(function () {
     return controller.get(req.uri);
   }, res);
@@ -52,7 +53,6 @@ function routes(router) {
   router.get('/:name', getUriFromReference);
   router.put('/:name', putUriFromReference);
   router.delete('/:name', deleteUriFromReference);
-
 }
 
 module.exports = routes;
