+++ lib/services/composer.js
@@ -15,10 +15,11 @@ const _ = require('lodash'),
  * @param {function|string} [filter='_ref']
  * @returns {Promise} - Resolves with composed component data
  */
-function resolveComponentReferences(data, locals, filter) {
-  const referenceObjects = references.listDeepObjects(data, filter || referenceProperty);
+function resolveComponentReferences(data, locals, filter = referenceProperty) {
+  const referenceObjects = references.listDeepObjects(data, filter);
 
   return bluebird.all(referenceObjects).each(function (referenceObject) {
+
     return components.get(referenceObject[referenceProperty], locals)
       .then(function (obj) {
         // the thing we got back might have its own references
