+++ lib/setup.js
@@ -4,6 +4,7 @@ const bluebird = require('bluebird'),
   express = require('express'),
   routes = require('./routes'),
   bootstrap = require('./bootstrap'),
+  render = require('./render'),
   htmlComposer = require('./html-composer');
 
 bluebird.config({
@@ -21,20 +22,21 @@ bluebird.config({
  * @returns {Promise}
  */
 module.exports = function (options) {
-  let app, engines, providers, router, sessionStore;
+  let app, engines, providers, router, sessionStore, renderers;
 
   options = options || {};
   app = options.app || express(); // use new express app if not passed in
   engines = options.engines; // will be undefined if not passed in
   providers = options.providers || [];
   sessionStore = options.sessionStore;
+  renderers = options.renderers || {}; // TODO: DOCUMENT THIS
 
   // init the router
   router = routes(app, providers, sessionStore);
 
-  // if engines were passed in, send them to the composer
-  if (engines) {
-    htmlComposer.addEngines(engines);
+  // if engines were passed in, send them to the renderer
+  if (renderers) {
+    render.registerRenderers(renderers);
   }
 
   // look for bootstraps in components
