+++ package.json
@@ -16,6 +16,7 @@
   "author": "New York Media",
   "license": "MIT",
   "dependencies": {
+    "amphora-html": "file:///Users/jonwinton/Sites/amphora-html",
     "basic-auth": "^1.0.4",
     "bluebird": "^3.0.0",
     "body-parser": "^1.12",
